use our own postrotate script for rsyslogd
The original script is like this:

    #!/bin/sh

    if [ -d /run/systemd/system ]; then
        systemctl kill -s HUP rsyslog.service
    else
        invoke-rc.d rsyslog rotate > /dev/null
    fi

Which just serves to send kill -HUP using the appropriate expected layer
(SystemD or Upstart). In our docker container we have none of those,
it was running the Upstart version (invoke-rc.d) which wasn't working.

Here we replace it by a simple script that runs the kill -HUP directly
which is simple and works.